[autoscaler]reconstruct upon Monorepo - Phase 1
