+++ src/frontend/packages/core/src/features/applications/application/application-tabs-base/tabs/autoscaler-tab/autoscaler-tab.component.html
