+++ src/frontend/packages/core/src/features/applications/autoscaler-metric-chart-page/combo-chart/combo-series-vertical.component.ts
