+++ src/frontend/packages/core/src/features/applications/autoscaler-metric-chart-page/combo-chart/index.ts
