+++ src/frontend/packages/core/src/features/applications/autoscaler-scale-history-page/autoscaler-scale-history-page.component.html
