+++ src/frontend/packages/core/src/features/applications/autoscaler-scale-history-page/autoscaler-scale-history-page.component.scss
@@ -4,7 +4,7 @@
     flex-grow: 1;
   }
 
-  .cdk-column-actor_name {
+  .cdk-column-actor-name {
     align-items: center;
     display: flex;
     flex-basis: 100px;
