+++ src/frontend/packages/core/src/features/applications/edit-autoscaler-policy/edit-autoscaler-policy.component.scss
