+++ src/frontend/packages/core/src/shared/components/cards/card-autoscaler-default/card-autoscaler-default.component.spec.ts
