+++ src/frontend/packages/core/src/shared/components/cards/card-autoscaler-default/card-autoscaler-default.component.ts
@@ -2,7 +2,7 @@ import { Component, ElementRef, OnDestroy, OnInit, ViewChild } from '@angular/co
 import { Observable } from 'rxjs';
 import { map, first, filter } from 'rxjs/operators';
 import { Store } from '@ngrx/store';
-import { AppState } from '../../../../store/app-state';
+import { AppState } from '../../../../../../store/src/app-state';
 import { ApplicationService } from '../../../../features/applications/application.service';
 import { MatSnackBar, MatSnackBarRef, SimpleSnackBar } from '@angular/material';
 import { EntityService } from '../../../../core/entity-service';
@@ -10,11 +10,11 @@ import { EntityServiceFactory } from '../../../../core/entity-service-factory.se
 import {
   entityFactory,
   appAutoscalerPolicySchemaKey,
-} from '../../../../store/helpers/entity-factory';
-import { GetAppAutoscalerPolicyAction, UpdateAppAutoscalerPolicyAction } from '../../../../store/actions/app-autoscaler.actions';
-import { AppAutoscalerPolicy } from '../../../../store/types/app-autoscaler.types';
-import { selectUpdateInfo } from '../../../../store/selectors/api.selectors';
-import { ActionState } from '../../../../store/reducers/api-request-reducer/types';
+} from '../../../../../../store/src/helpers/entity-factory';
+import { GetAppAutoscalerPolicyAction, UpdateAppAutoscalerPolicyAction } from '../../../../../../store/src/actions/app-autoscaler.actions';
+import { AppAutoscalerPolicy } from '../../../../../../store/src/types/app-autoscaler.types';
+import { selectUpdateInfo } from '../../../../../../store/src/selectors/api.selectors';
+import { ActionState } from '../../../../../../store/src/reducers/api-request-reducer/types';
 import { CardStatus } from '../../../shared.types';
 
 @Component({
