+++ src/frontend/packages/core/src/shared/components/list/list-types/app-autoscaler-event/cf-app-autoscaler-events-config.service.spec.ts
@@ -5,12 +5,12 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { CoreModule } from '../../../../../core/core.module';
 import { EntityServiceFactory } from '../../../../../core/entity-service-factory.service';
 import { ApplicationsModule } from '../../../../../features/applications/applications.module';
-import { GetApplication } from '../../../../../store/actions/application.actions';
-import { applicationSchemaKey, entityFactory } from '../../../../../store/helpers/entity-factory';
-import { endpointStoreNames } from '../../../../../store/types/endpoint.types';
-import { generateTestApplicationServiceProvider } from '../../../../../test-framework/application-service-helper';
-import { generateTestEntityServiceProvider } from '../../../../../test-framework/entity-service.helper';
-import { createBasicStoreModule, getInitialTestStoreState } from '../../../../../test-framework/store-test-helper';
+import { GetApplication } from '../../../../../../../store/src/actions/application.actions';
+import { applicationSchemaKey, entityFactory } from '../../../../../../../store/src/helpers/entity-factory';
+import { endpointStoreNames } from '../../../../../../../store/src/types/endpoint.types';
+import { generateTestApplicationServiceProvider } from '../../../../../../test-framework/application-service-helper';
+import { generateTestEntityServiceProvider } from '../../../../../../test-framework/entity-service.helper';
+import { createBasicStoreModule, getInitialTestStoreState } from '../../../../../../test-framework/store-test-helper';
 import { SharedModule } from '../../../../shared.module';
 import { CfAppAutoscalerEventsConfigService } from './cf-app-autoscaler-events-config.service';
