+++ src/frontend/packages/core/src/shared/components/list/list-types/app-autoscaler-event/table-cell-autoscaler-event-action/table-cell-autoscaler-event-action.component.ts
