+++ src/frontend/packages/core/src/shared/components/list/list-types/app-autoscaler-event/table-cell-autoscaler-event-change/table-cell-autoscaler-event-change.component.scss
