+++ src/frontend/packages/core/src/shared/components/list/list-types/app-autoscaler-event/table-cell-autoscaler-event-error/table-cell-autoscaler-event-error.component.html
