+++ src/frontend/packages/core/src/shared/components/list/list-types/app-autoscaler-event/table-cell-autoscaler-event-status/table-cell-autoscaler-event-status.component.spec.ts
