+++ src/frontend/packages/core/src/shared/components/list/list-types/app-autoscaler-event/table-cell-autoscaler-event-timestamp/table-cell-autoscaler-event-timestamp.component.ts
