+++ src/frontend/packages/core/src/shared/components/list/list-types/app-autoscaler-event/table-cell-autoscaler-event-type/table-cell-autoscaler-event-type.component.spec.ts
