+++ src/frontend/packages/store/src/helpers/autoscaler-helpers.spec.ts
