+++ src/frontend/packages/store/src/types/app-autoscaler.types.ts
